+++ app/src/main/java/woowacourse/omok/MainActivity.kt
@@ -5,6 +5,8 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.omok.data.Player
 import woowacourse.omok.data.Room
+import woowacourse.omok.dbHelper.OmokPlayerDbHelper
+import woowacourse.omok.dbHelper.OmokRoomDbHelper
 
 class MainActivity : AppCompatActivity() {
     private val playerNames: List<Player> =
