+++ app/src/main/java/woowacourse/omok/dbHelper/OmokPlayerConstract.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.dbHelper
 
 object OmokPlayerConstract {
     const val TABLE_NAME_PLAYER = "players"
